+++ CHANGES.rst
@@ -21,6 +21,13 @@ Features
 - Add a ``missing_msg`` argument to ``SchemaNode`` that specifies the error
   message to be used when the node is required and missing
 
+Bug Fixes
+---------
+
+- Fix an issue where the ``flatten()`` method produces an invalid name
+  (ex: "answer.0.") for the type "Sequence".  See
+  https://github.com/Pylons/colander/issues/179
+
 
 1.0 (2014-11-26)
 ----------------